requirements.txt
elasticsearch
sentence-transformers
numpy
pandas
faiss-cpu
matplotlib
jupyter
scikit-learn
transformers
beautifulsoup4
lxml
bs4